WIP - Implement ITM Egress Tunnel Tables

+++ itm/itm-api/src/main/java/org/opendaylight/genius/itm/utils/DpnTepInterfaceInfo.java
@@ -17,7 +17,7 @@ public interface DpnTepInterfaceInfo {
 
     Class<? extends TunnelTypeBase> getTunnelType();
 
-    long getGroupId();
+    int getGroupId();
 
     String getTunnelName();
