+++ itm/itm-impl/src/main/java/org/opendaylight/genius/itm/confighelpers/ItmInternalTunnelAddWorker.java
@@ -345,7 +345,7 @@ public final class ItmInternalTunnelAddWorker {
 
         //ITM TEP INTERFACE set the group Id here ..later
         Integer groupId = directTunnelUtils.allocateId(ITMConstants.ITM_IDPOOL_NAME, srcDpnId.toString());
-        dpnsTepsBuilder.setGroupId(groupId.longValue());
+        dpnsTepsBuilder.setGroupId(groupId);
         RemoteDpnsBuilder remoteDpn = new RemoteDpnsBuilder();
         remoteDpn.withKey(new RemoteDpnsKey(dstDpnId));
         remoteDpn.setDestinationDpnId(dstDpnId);
